+++ satella/coding/structures/__init__.py
@@ -5,7 +5,7 @@ from .hashable_objects import HashableWrapper
 from .heaps import Heap, SetHeap, TimeBasedHeap, TimeBasedSetHeap
 from .immutable import Immutable, frozendict
 from .mixins import OmniHashableMixin, ReprableMixin, StrEqHashableMixin, ComparableIntEnum, \
-    HashableIntEnum, ComparableAndHashableBy
+    HashableIntEnum, ComparableAndHashableBy, ComparableAndHashableByInt
 from .proxy import Proxy
 from .ranking import Ranking
 from .singleton import Singleton, SingletonWithRegardsTo
@@ -17,6 +17,7 @@ __all__ = [
     'KeyAwareDefaultDict',
     'Proxy',
     'ReprableMixin',
+    'ComparableAndHashableByInt',
     'StrEqHashableMixin',
     'ComparableIntEnum',
     'ComparableAndHashableBy',