+++ satella/coding/transforms/__init__.py
@@ -7,9 +7,26 @@ from satella.coding.decorators import for_argument
 from .merger import merge_series
 
 __all__ = ['stringify', 'split_shuffle_and_join', 'one_tuple',
-           'merge_series', 'pad_to_multiple_of_length']
+           'merge_series', 'pad_to_multiple_of_length', 'clip']
 
 T = tp.TypeVar('T')
+Number = tp.Union[int, float]
+
+
+def clip(v: Number, min: Number, max: Number) -> Number:
+    """
+    Clip v so it conforms to min <= v <= max
+
+    :param v: value to clip
+    :param min: minimum
+    :param max: maximum
+    :return: clipped value
+    """
+    if v < min:
+        return min
+    if v > max:
+        return max
+    return v
 
 
 @for_argument(list)