+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.10.0
 
 * added `satella.coding.predicates`
+* added `ComparableAndHashableByInt`