+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 
 * added `satella.coding.predicates`
 * added `ComparableAndHashableByInt`
+* added slots to some mix-ins
+