+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 * `for_argument` can now accept strings
 * changed typing for `for_argument`
 * added `source_to_function`
+* added `clip`