+++ CHANGELOG.md
@@ -7,4 +7,4 @@
 * added predicate `has_keys`, deprecated
   existing `has_keys`
 * changed ordering of arguments to attribute and item
-
+* `enumerate` -> `smart_enumerate`