+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# v2.10.1
+# v2.11.0
 
 * `for_argument` can now accept strings
 * changed typing for `for_argument`
@@ -7,4 +7,5 @@
 * added predicate `has_keys`, deprecated
   existing `has_keys`
 * changed ordering of arguments to attribute and item
+* overhauled `predicates`
 * `enumerate` -> `smart_enumerate`